requirements.txt
torch==1.6.0
streamlit==0.67.1
numpy==1.19.1
torchvision==0.7.0
onnx==1.7.0
onnx_caffe2==1.0.0
Pillow==7.2.0
